Extended gridData to work with env/trait data
gridData now takes a type argument:
- type == "geo" is equivalent to the prior function - for gridding
geographical coordinates.
- type == "env" is associated with the site stats of dist_data and bins
"sites" in multidimensional environmental space. It requires the
argument env_binsizes, which takes a vector of length equal to the
number of site stats, where each number is the desired bin size for each
site stat. It returns an object of type "env_grid".
- type == "trait" is associated with the species stats of dist_data and
bins "sites" in multidimensional trait space. It requires the argument
trait_binsizes, which takes a vector of length equal to the number of
variables in species_stats, where each number is the desired bin size
for each species_stats variable. It returns an object of type
"trait_grid"